Large chunk of modifications / refactoring
Modules
=======
Modules (e.g. Fire, Grass) are derived from "Module" and can be
registered to be responsible for a certain type of State, or for running
every year (e.g. fire module). The model determines which modules are
enabled and creates them on the fly. Now both disturbance modules and
state-change modules (e.g. grass) share the same structure.

Batches
=======
There are now two types of batches: one for DNN processing, one for
other processing. "Other" batches are not sent to the DNN thread, but to
worker threads of the main model (a "module" is linked to a batch and
each batch is filled with cells that should be treated by the same
module / model; in addition, the linked module does then the processing)

"Grass" Module
==============
This is just a (minimal) demo module; it defines 5 grass states (with a
transition matrix); burned px have a prob to turn to grass land, and
from one of the grass states there is again a prob that a px turns into
forest again.

States
======
States have now a type ("landcover type"). Modules can be responsible
for certain types; for example, forest states can be processed via the
main DNN, while grass states could be processed by a grass module (with
transition matrices).