Multiple DNN instances in SVD
Refactoring of the DNN code that allows to run multiple instances of a
DNN in parallel. Each DNN instance sets up the same network and
settings. On execution of a batch different DNN instances are selected
(cycling through all available DNNs).
Fixed a crash when stats are collected before BatchManager is available.