Compilation and execution on Linux
Update to Tensorflow 2.x -> this uses now a precompiled version of
libtensorflow_cc (from https://github.com/ika-rwth-aachen/
libtensorflow_cc). Needed to change some paths, but otherwise went
unexpectely smooth.

Update to C++ 14 (from C++ 11) due to compilation errors with Eigen. I
suspect that some of the follow-up errors were due to this (mostly
missing includes of std::string and such).

Locale: switched now SVD to the "C" locale, so that e.g. atof("0.9")
works also on a German system. The goal is here to have input files that
are language-agnostic, i.e. that work everywhere when a project is
transferred.