Major change in the representation of cells/grids
Adding of a state/residence time history (3 previous steps).

Previously, the main container for cells on the landscape was a
Grid<Cell>, thus requiring memory also for cells not forested.

Now with the Europe project upcoming, it turned out that this is
particularly wasteful. For Europe only 5-10% of cells are actually
forested. Having 64 bytes for every cell would amount to 160GB of memory
requirement. Now we only need four bytes per cell (an integer index to
the vector<Cells>) and the estimate is ~20GB of RAM (3e9 * (4 +
0.05*64)).

The single int is part encapsulated in a "GridCell" class, which is a
convenience and allows to access the cell.

The change required moderate changes on many different places.