bringing SVD to Lorien
multiple changes to run SVD on Lorien - partly to compile (with different
versions of Tensorflow. It compiles with 2.9.3, but not with 2.11).
some changes to allow INT2S / INT2 for TIFs to load data.

played around with using multiple GPUs, but that does not seem to work,
partly due to a bug in Tensorflow (e.g. https://github.com/tensorflow/
tensorflow/issues/1888#issuecomment-369407156). Maybe this is resolved in
future versions...

With the changes it was possible to load Europe into SVD (yeah).
However, there are crashes (SIGSEGV) deep within Tensorflow. Need to be
investigated further.