wind/beetle interaction
simple interaction of wind and beetles. After wind events, a bark beetle
outbreak starts at the respective pixel with a certain chance. Right now
the sequence of modules is determinsitic and depends on the sequence in
the config file.

Added a type property (as string) to modules. This allows to look for a
module of a certain type.