Switch from QtVisualization to QtGraphs
Since QtVisualization is deprecated soon, and there were a lot of
troubles with hanging SVD on Linux / remote desktop / Qt 6, I decided to
switch to QtVisualization.
This is the first batch of changes, showing that it is actually working.
Not everything is working right now, but the foundation is here!